[Feature #163248472]Add find by username function.
